requirements.txt
openai
transformers
opencv-python
neo4j
streamlit
pillow
python-dotenv
dotenv
torch
torchvision
torchaudio